refector/#19: combineapi.py 제거, CORS 설정 추가
