+++ requirements.txt
@@ -1,22 +1,38 @@
+annotated-types==0.7.0
+anyio==4.10.0
 asgiref==3.9.1
 attrs==25.3.0
 certifi==2025.8.3
 charset-normalizer==3.4.2
+colorama==0.4.6
+distro==1.9.0
 Django==5.2.5
+django-cors-headers==4.7.0
 djangorestframework==3.16.1
 drf-spectacular==0.28.0
 drf-spectacular-sidecar==2025.8.1
+h11==0.16.0
+httpcore==1.0.9
+httpx==0.28.1
 idna==3.10
 inflection==0.5.1
+jiter==0.10.0
 jsonschema==4.25.0
 jsonschema-specifications==2025.4.1
+openai==1.99.6
 pillow==11.3.0
+pydantic==2.11.7
+pydantic_core==2.33.2
 python-dotenv==1.1.1
 PyYAML==6.0.2
 referencing==0.36.2
 requests==2.32.4
 rpds-py==0.27.0
+sniffio==1.3.1
 sqlparse==0.5.3
+tqdm==4.67.1
+typing-inspection==0.4.1
+typing_extensions==4.14.1
 tzdata==2025.2
 uritemplate==4.2.0
 urllib3==2.5.0
